Moved some common code to the common folder
