+++ src/components/common/images/FactionLogo.tsx
