+++ src/components/common/images/UnitProfilePicture.tsx
