+++ src/constants/wanderers.ts
